adding more stuff, not fully functional

+++ app/src/main/java/com/example/levelup/dataParser/Vector3.java
@@ -5,7 +5,7 @@ public class Vector3 {
     private double y;
     private double z;
 
-    public Vector3(float x, float y, float z) {
+    public Vector3(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
